Fixed(?) Spread thread-safety issue in src/rsb/transport/spread/SpreadConnection.{h,cpp}
At least for sending messages, the Spread client library does not seem
to be thread-safe on win32 (despite what the documentation says).

This fix seems to work but has only been verified experimentally. In
particular, it is not clear, whether all other combinations of Spread
client API functions are thread-safe.

fixes #2386

* src/rsb/transport/spread/SpreadConnection.{h,cpp} (header): updated
  copyright
  (SpreadConnection::mutex): new member; for critical section in send
  on win32
  (SpreadConnection::send): turned Spread client API calls into
  critical section on win32